web/requirements.txt
Flask
flask_restful
pymongo
python-dotenv
dnspython
flask_cors
pillow
numpy